requirements.txt
numpy
scipy
pandas
pyfunctional
fastcluster
biopython
PySide
ete3